chore(scripts): move sprite generators into scripts/sprites/
Group the submit footer sprite generators and the analysis utility under their own subdirectory so the top-level scripts/ folder isn't dominated by gen_*_sprite.py files. Update the in-script asset path resolution to walk one extra parent level.
